refactor(drive-abci): split document query v1 routing and dispatchers into files
Pure motion, no behavior change: mod.rs was 1748 lines holding the
shape-routing table, six per-route dispatchers, and their wire-shape
helpers. Now mod.rs keeps the handler entry (query_documents_v1),
RoutingDecision, and the shared not_yet_implemented helper; routing.rs
holds validate_and_route + the offset gate + the test probe; and
dispatch/ holds one file per RoutingDecision arm (documents, count,
sum, average, ranked, having), each with its own impl block and its
route-local helpers, with the entry-shape helpers shared by the
ranked + having pair in dispatch/mod.rs.
