+++ packages/rs-drive-abci/src/query/document_query/v1/dispatch/count.rs
@@ -1,0 +1,213 @@
+//! `RoutingDecision::Count` — the count surfaces (aggregate and
+//! per-group entries).
+
+use super::super::not_yet_implemented;
+use crate::error::query::QueryError;
+use crate::error::Error;
+use crate::platform_types::platform::Platform;
+use crate::platform_types::platform_state::PlatformState;
+use crate::query::response_metadata::CheckpointUsed;
+use crate::query::QueryValidationResult;
+use dapi_grpc::platform::v0::get_documents_request::get_documents_request_v1::Start as RequestV1Start;
+use dapi_grpc::platform::v0::get_documents_response::get_documents_response_v1::{
+    count_results, result_data, CountEntries, CountEntry, CountResults, ResultData,
+};
+use dapi_grpc::platform::v0::get_documents_response::{
+    get_documents_response_v1, GetDocumentsResponseV1,
+};
+use dpp::check_validation_result_with_data;
+use dpp::data_contract::accessors::v0::DataContractV0Getters;
+use dpp::identifier::Identifier;
+use dpp::validation::ValidationResult;
+use dpp::version::PlatformVersion;
+use drive::error::query::QuerySyntaxError;
+use drive::query::{
+    CountMode, DocumentCountRequest, DocumentCountResponse, OrderClause, SplitCountEntry,
+    WhereClause,
+};
+use drive::util::grove_operations::GroveDBToUse;
+
+impl<C> Platform<C> {
+    /// Forward a `select = COUNT` request to drive's count
+    /// dispatcher. `mode` is the SQL-shape contract derived from
+    /// `(select, group_by, where)` by `validate_and_route`; drive
+    /// uses it to pick the executor strategy and decide whether to
+    /// collapse the response to a single aggregate or return per-
+    /// group entries. The wire response is `GetDocumentsResponseV1`
+    /// with the inner `ResultData.counts` variant for non-proof
+    /// results.
+    #[allow(clippy::too_many_arguments)]
+    pub(in crate::query::document_query::v1) fn dispatch_count_v1(
+        &self,
+        data_contract_id: Vec<u8>,
+        document_type_name: String,
+        where_clauses: Vec<WhereClause>,
+        order_clauses: Vec<OrderClause>,
+        limit: Option<u32>,
+        start: Option<RequestV1Start>,
+        prove: bool,
+        mode: CountMode,
+        platform_state: &PlatformState,
+        platform_version: &PlatformVersion,
+    ) -> Result<QueryValidationResult<GetDocumentsResponseV1>, Error> {
+        if start.is_some() {
+            return Ok(QueryValidationResult::new_with_error(not_yet_implemented(
+                "start_after / start_at with SELECT COUNT (paginate by narrowing the \
+                 range clause itself)",
+            )));
+        }
+
+        let contract_id: Identifier =
+            check_validation_result_with_data!(data_contract_id.try_into().map_err(|_| {
+                QueryError::InvalidArgument(
+                    "id must be a valid identifier (32 bytes long)".to_string(),
+                )
+            }));
+
+        let (_, contract_fetch_info) = self.drive.get_contract_with_fetch_info_and_fee(
+            contract_id.to_buffer(),
+            None,
+            true,
+            None,
+            platform_version,
+        )?;
+        let contract_fetch_info = check_validation_result_with_data!(contract_fetch_info.ok_or(
+            QueryError::Query(QuerySyntaxError::DataContractNotFound(
+                "contract not found when querying from value with contract info",
+            ))
+        ));
+        let contract_ref = &contract_fetch_info.contract;
+        let document_type = check_validation_result_with_data!(contract_ref
+            .document_type_for_name(document_type_name.as_str())
+            .map_err(|_| QueryError::InvalidArgument(format!(
+                "document type {} not found for contract {}",
+                document_type_name, contract_id
+            ))));
+
+        let drive_request = DocumentCountRequest {
+            contract: contract_ref,
+            document_type,
+            where_clauses,
+            order_clauses,
+            mode,
+            limit,
+            prove,
+            drive_config: &self.config.drive,
+        };
+        let drive_response =
+            match self
+                .drive
+                .execute_document_count_request(drive_request, None, platform_version)
+            {
+                Ok(r) => r,
+                Err(drive::error::Error::Query(qe)) => {
+                    return Ok(QueryValidationResult::new_with_error(QueryError::Query(qe)));
+                }
+                Err(e) => return Err(e.into()),
+            };
+
+        let response = match drive_response {
+            DocumentCountResponse::Aggregate(count) => GetDocumentsResponseV1 {
+                result: Some(get_documents_response_v1::Result::Data(ResultData {
+                    variant: Some(result_data::Variant::Counts(CountResults {
+                        variant: Some(count_results::Variant::AggregateCount(count)),
+                    })),
+                })),
+                metadata: Some(self.response_metadata_v0(platform_state, CheckpointUsed::Current)),
+            },
+            DocumentCountResponse::Entries(entries) => {
+                if mode.is_aggregate() {
+                    // `select=COUNT, group_by=[]` against a request
+                    // that drove a PerInValue execution (In + no
+                    // range + no prove). Sum entries into a single
+                    // aggregate before emission. `checked_add`
+                    // surfaces u64 overflow as a typed
+                    // `QuerySyntaxError::Unsupported`; realistic
+                    // ceiling is `|In| × max_per-branch-count` (well
+                    // under u64), so triggering this path requires
+                    // either a misconfigured count tree or an
+                    // executor bug.
+                    let mut total: u64 = 0;
+                    let mut overflow = false;
+                    for e in &entries {
+                        // `count.unwrap_or(0)` here is safe: this
+                        // arm is server-side, summing entries the
+                        // executor emitted. Executor never emits
+                        // `None` (that's an SDK-side
+                        // synthesis-for-missing concept). The
+                        // `unwrap_or(0)` is a belt-and-suspenders
+                        // guard against any future executor that
+                        // forgets the contract.
+                        match total.checked_add(e.count.unwrap_or(0)) {
+                            Some(t) => total = t,
+                            None => {
+                                overflow = true;
+                                break;
+                            }
+                        }
+                    }
+                    if overflow {
+                        return Ok(QueryValidationResult::new_with_error(QueryError::Query(
+                            QuerySyntaxError::Unsupported(
+                                "aggregate COUNT across In branches overflows u64 — \
+                                 narrow the In set or query branches individually"
+                                    .to_string(),
+                            ),
+                        )));
+                    }
+                    GetDocumentsResponseV1 {
+                        result: Some(get_documents_response_v1::Result::Data(ResultData {
+                            variant: Some(result_data::Variant::Counts(CountResults {
+                                variant: Some(count_results::Variant::AggregateCount(total)),
+                            })),
+                        })),
+                        metadata: Some(
+                            self.response_metadata_v0(platform_state, CheckpointUsed::Current),
+                        ),
+                    }
+                } else {
+                    GetDocumentsResponseV1 {
+                        result: Some(get_documents_response_v1::Result::Data(ResultData {
+                            variant: Some(result_data::Variant::Counts(CountResults {
+                                variant: Some(count_results::Variant::Entries(CountEntries {
+                                    entries: entries.into_iter().map(into_v1_entry).collect(),
+                                })),
+                            })),
+                        })),
+                        metadata: Some(
+                            self.response_metadata_v0(platform_state, CheckpointUsed::Current),
+                        ),
+                    }
+                }
+            }
+            DocumentCountResponse::Proof(proof_bytes) => {
+                let (grovedb_used, proof) =
+                    self.response_proof_v0(platform_state, proof_bytes, GroveDBToUse::Current)?;
+                GetDocumentsResponseV1 {
+                    result: Some(get_documents_response_v1::Result::Proof(proof)),
+                    metadata: Some(self.response_metadata_v0(platform_state, grovedb_used)),
+                }
+            }
+        };
+
+        Ok(QueryValidationResult::new_with_data(response))
+    }
+}
+
+fn into_v1_entry(e: SplitCountEntry) -> CountEntry {
+    CountEntry {
+        in_key: e.in_key,
+        key: e.key,
+        // The wire `count` is `uint64`, so it can only carry
+        // `Some(_)`. Server-side never emits `None` entries to
+        // begin with — `None` is the SDK-side synthesis signal for
+        // "caller's In array contained a value the proof was
+        // silent on," and that decision lives client-side because
+        // the wire never has the caller's full In array context.
+        // `unwrap_or(0)` is defense-in-depth: a future executor
+        // bug emitting `None` shouldn't crash the response path,
+        // it should round to zero on the wire (matching the
+        // proto's `uint64` default).
+        count: e.count.unwrap_or(0),
+    }
+}
